change to float

+++ src/com/melody/utils/CoordUtils.java
@@ -2,9 +2,9 @@ package com.melody.utils;
 
 public class CoordUtils {
 	
-	public final static boolean pointInRect(int coordX, int coordY, int x, int y, int width, int height) {
-		int mathx = x + width - 1;
-        int mathy = y + height - 1;
+	public final static boolean pointInRect(float coordX, float coordY, float x, float y, float width, float height) {
+		float mathx = x + width - 1;
+		float mathy = y + height - 1;
         return coordX >= x && coordX <= mathx && coordY >= y && coordY <= mathy;
 	}
